(SIMP-1190) Idempotency fix fo Exec[bootstrap_ldap]
This explicitly sets the ownership on the bootstrap lockfile created by
`Exec[bootstrap_ldap]` via its command in order to remain idempotent.
Previously the file was created as the user executing Puppet (usually
`root:root`) which conflicted with the recursively-managed directory in
which it is contained.  Because the directory was created first, the
`File` did adjust the group ownership of the lock file, but instead
would correct it on a subsiquent run-- a non-idempotent action.