(SIMP-2992) Work around for broken LDAP libs
The client LDAP libraries are broken in EL6 such that, if any 128 bit
ciphers are present in the cipher string, it will fall back to SSF=128
for connecting to an LDAP server.

This is unacceptable to the SIMP default configuration (SSF=256).

Also, fixed a bug in a template with regards to variables with prefaced
underscores.

SIMP-2992 #close Update for OpenLDAP Client