(SIMP-3894) Allow custom GIDs in default.ldif
Allow users to set the 'users' and 'administrators' GID values in the
'default.ldif' file.

These were *not* exposed at a higher level to prevent needing to make
breaking changes in the near future and given that none of the other
settings were exposed at that level.

If this is needed, we will need to create another change in the future.

SIMP-4131 #close
SIMP-4132 #close